opencv_engine: Rename submodule back to engine
