+++ opencv_engine/engine.py
@@ -8,11 +8,8 @@
 # http://www.opensource.org/licenses/mit-license
 # Copyright (c) 2014 globo.com timehome@corp.globo.com
 
-try:
-    import cv
-except ImportError:
-    import cv2.cv as cv
-
+import cv2
+import numpy as np
 from colour import Color
 from pexif import ExifSegment, JpegFile
 from thumbor.engines import BaseEngine
@@ -24,21 +21,31 @@ try:
 except ImportError:
     FILTERS_AVAILABLE = False
 
-FORMATS = {".jpg": "JPEG", ".jpeg": "JPEG", ".gif": "GIF", ".png": "PNG"}
+FORMATS = {
+    ".jpg": "JPEG",
+    ".jpeg": "JPEG",
+    ".gif": "GIF",
+    ".png": "PNG",
+    ".webp": "WEBP",
+}
 
 
 class Engine(BaseEngine):
     @property
     def image_depth(self):
         if self.image is None:
-            return 8
-        return cv.GetImage(self.image).depth
+            return np.uint8
+        return self.image.dtype
 
     @property
     def image_channels(self):
         if self.image is None:
             return 3
-        return self.image.channels
+        # if the image is grayscale
+        try:
+            return self.image.shape[2]
+        except IndexError:
+            return 1
 
     @classmethod
     def parse_hex_color(cls, color):
@@ -49,15 +56,16 @@ class Engine(BaseEngine):
             return None
 
     def gen_image(self, size, color_value):
-        img0 = cv.CreateImage(size, self.image_depth, self.image_channels)
         if color_value == "transparent":
             color = (255, 255, 255, 255)
+            img = np.zeros((size[1], size[0], 4), self.image_depth)
         else:
+            img = np.zeros((size[1], size[0], self.image_channels), self.image_depth)
             color = self.parse_hex_color(color_value)
             if not color:
                 raise ValueError("Color %s is not valid." % color_value)
-        cv.Set(img0, color)
-        return img0
+        img[:] = color
+        return img
 
     def create_image(self, buffer):
         # FIXME: opencv doesn't support gifs, even worse, the library
@@ -69,11 +77,9 @@ class Engine(BaseEngine):
         except KeyError:
             pass
 
-        imagefiledata = cv.CreateMatHeader(1, len(buffer), cv.CV_8UC1)
-        cv.SetData(imagefiledata, buffer, len(buffer))
-        img0 = cv.DecodeImageM(imagefiledata, cv.CV_LOAD_IMAGE_UNCHANGED)
-
+        img = cv2.imdecode(np.frombuffer(buffer, np.uint8), -1)
         if FORMATS[self.extension] == "JPEG":
+            self.exif = None
             try:
                 info = JpegFile.fromString(buffer).get_exif()
                 if info:
@@ -81,64 +87,60 @@ class Engine(BaseEngine):
                     self.exif_marker = info.marker
             except Exception:
                 pass
-
-        return img0
+        return img
 
     @property
     def size(self):
-        return cv.GetSize(self.image)
+        return self.image.shape[1], self.image.shape[0]
 
     def normalize(self):
         pass
 
     def resize(self, width, height):
-        thumbnail = cv.CreateImage(
-            (int(round(width, 0)), int(round(height, 0))),
-            self.image_depth,
-            self.image_channels,
-        )
-        cv.Resize(self.image, thumbnail, cv.CV_INTER_AREA)
-        self.image = thumbnail
+        r = height / self.size[1]
+        width = int(self.size[0] * r)
+        dim = (int(round(width, 0)), int(round(height, 0)))
+        self.image = cv2.resize(self.image, dim, interpolation=cv2.INTER_AREA)
 
     def crop(self, left, top, right, bottom):
-        new_width = right - left
-        new_height = bottom - top
-        cropped = cv.CreateImage(
-            (new_width, new_height), self.image_depth, self.image_channels
-        )
-        src_region = cv.GetSubRect(self.image, (left, top, new_width, new_height))
-        cv.Copy(src_region, cropped)
-
-        self.image = cropped
+        self.image = self.image[top:bottom, left:right]
 
     def rotate(self, degrees):
-        if degrees > 180:
-            # Flip around both axes
-            cv.Flip(self.image, None, -1)
-            degrees = degrees - 180
-
-        img = self.image
-        size = cv.GetSize(img)
-
-        if degrees / 90 % 2:
-            new_size = (size[1], size[0])
-            center = ((size[0] - 1) * 0.5, (size[0] - 1) * 0.5)
+        # see http://stackoverflow.com/a/23990392
+        if degrees == 90:
+            self.image = cv2.transpose(self.image)
+            cv2.flip(self.image, 0, self.image)
+        elif degrees == 180:
+            cv2.flip(self.image, -1, self.image)
+        elif degrees == 270:
+            self.image = cv2.transpose(self.image)
+            cv2.flip(self.image, 1, self.image)
         else:
-            new_size = size
-            center = ((size[0] - 1) * 0.5, (size[1] - 1) * 0.5)
+            # see http://stackoverflow.com/a/37347070
+            # one pixel glitch seems to happen with 90/180/270
+            # degrees pictures in this algorithm if you check
+            # the typical github.com/recurser/exif-orientation-examples
+            # but the above transpose/flip algorithm is working fine
+            # for those cases already
+            width, height = self.size
+            image_center = (width / 2, height / 2)
+            rot_mat = cv2.getRotationMatrix2D(image_center, degrees, 1.0)
 
-        mapMatrix = cv.CreateMat(2, 3, cv.CV_64F)
-        cv.GetRotationMatrix2D(center, degrees, 1.0, mapMatrix)
-        dst = cv.CreateImage(new_size, self.image_depth, self.image_channels)
-        cv.SetZero(dst)
-        cv.WarpAffine(img, dst, mapMatrix)
-        self.image = dst
+            abs_cos = abs(rot_mat[0, 0])
+            abs_sin = abs(rot_mat[0, 1])
+            bound_w = int((height * abs_sin) + (width * abs_cos))
+            bound_h = int((height * abs_cos) + (width * abs_sin))
+
+            rot_mat[0, 2] += (bound_w / 2) - image_center[0]
+            rot_mat[1, 2] += (bound_h / 2) - image_center[1]
+
+            self.image = cv2.warpAffine(self.image, rot_mat, (bound_w, bound_h))
 
     def flip_vertically(self):
-        cv.Flip(self.image, None, 1)
+        self.image = np.flipud(self.image)
 
     def flip_horizontally(self):
-        cv.Flip(self.image, None, 0)
+        self.image = np.fliplr(self.image)
 
     def read(self, extension=None, quality=None):
         if quality is None:
@@ -148,15 +150,22 @@ class Engine(BaseEngine):
         extension = extension or self.extension
         try:
             if FORMATS[extension] == "JPEG":
-                options = [cv.CV_IMWRITE_JPEG_QUALITY, quality]
+                options = [cv2.IMWRITE_JPEG_QUALITY, quality]
         except KeyError:
             # default is JPEG so
-            options = [cv.CV_IMWRITE_JPEG_QUALITY, quality]
+            options = [cv2.IMWRITE_JPEG_QUALITY, quality]
 
-        data = cv.EncodeImage(extension, self.image, options or []).tostring()
+        try:
+            if FORMATS[extension] == "WEBP":
+                options = [cv2.IMWRITE_WEBP_QUALITY, quality]
+        except KeyError:
+            options = [cv2.IMWRITE_JPEG_QUALITY, quality]
+
+        success, buf = cv2.imencode(extension, self.image, options or [])
+        data = buf.tostring()
 
         if FORMATS[extension] == "JPEG" and self.context.config.PRESERVE_EXIF_INFO:
-            if hasattr(self, "exif"):
+            if hasattr(self, "exif") and self.exif != None:
                 img = JpegFile.fromString(data)
                 img._segments.insert(
                     0, ExifSegment(self.exif_marker, None, self.exif, "rw")
@@ -166,37 +175,45 @@ class Engine(BaseEngine):
         return data
 
     def set_image_data(self, data):
-        cv.SetData(self.image, data)
+        self.image = np.frombuffer(data, dtype=self.image.dtype).reshape(
+            self.image.shape
+        )
 
     def image_data_as_rgb(self, update_image=True):
-        # TODO: Handle other formats
         if self.image_channels == 4:
             mode = "BGRA"
         elif self.image_channels == 3:
             mode = "BGR"
         else:
             mode = "BGR"
-            rgb_copy = cv.CreateImage((self.image.width, self.image.height), 8, 3)
-            cv.CvtColor(self.image, rgb_copy, cv.CV_GRAY2BGR)
+            rgb_copy = np.zeros((self.size[1], self.size[0], 3), self.image.dtype)
+            cv2.cvtColor(self.image, cv2.COLOR_GRAY2BGR, rgb_copy)
             self.image = rgb_copy
         return mode, self.image.tostring()
 
     def draw_rectangle(self, x, y, width, height):
-        cv.Rectangle(
+        cv2.rectangle(
             self.image,
             (int(x), int(y)),
             (int(x + width), int(y + height)),
-            cv.Scalar(255, 255, 255, 1.0),
+            (255, 255, 255),
         )
 
-    def convert_to_grayscale(self):
-        if self.image_channels >= 3:
-            # FIXME: OpenCV does not support grayscale with alpha channel?
-            grayscaled = cv.CreateImage(
-                (self.image.width, self.image.height), self.image_depth, 1
-            )
-            cv.CvtColor(self.image, grayscaled, cv.CV_BGRA2GRAY)
-            self.image = grayscaled
+    def convert_to_grayscale(self, update_image=True, alpha=True):
+        image = None
+        if self.image_channels >= 3 and alpha:
+            image = cv2.cvtColor(self.image, cv2.COLOR_BGRA2GRAY)
+        elif self.image_channels >= 3:
+            image = cv2.cvtColor(self.image, cv2.COLOR_BGR2GRAY)
+        elif self.image_channels == 1:
+            # Already grayscale,
+            image = self.image
+        if update_image:
+            self.image = image
+        elif self.image_depth == np.uint16:
+            # Feature detector reqiures uint8 images
+            image = np.array(image, dtype="uint8")
+        return image
 
     def paste(self, other_engine, pos, merge=True):
         if merge and not FILTERS_AVAILABLE:
@@ -231,11 +248,9 @@ class Engine(BaseEngine):
 
     def enable_alpha(self):
         if self.image_channels < 4:
-            with_alpha = cv.CreateImage(
-                (self.image.width, self.image.height), self.image_depth, 4
-            )
+            with_alpha = np.zeros((self.size[1], self.size[0], 4), self.image.dtype)
             if self.image_channels == 3:
-                cv.CvtColor(self.image, with_alpha, cv.CV_BGR2BGRA)
+                cv2.cvtColor(self.image, cv2.COLOR_BGR2BGRA, with_alpha)
             else:
-                cv.CvtColor(self.image, with_alpha, cv.CV_GRAY2BGRA)
+                cv2.cvtColor(self.image, cv2.COLOR_GRAY2BGRA, with_alpha)
             self.image = with_alpha
